pyproject.toml
[project]
name = "ksef-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = []

[dependency-groups]
dev = [
    "alembic>=1.18.4",
    "alembic-utils>=0.8.8",
    "fastapi[standard]>=0.135.1",
    "httpx>=0.28.1",
    "psycopg[binary]>=3.3.3",
    "pydantic-settings>=2.13.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.2",
    "python-keycloak>=7.1.1",
    "requests>=2.32.5",
    "ruff>=0.15.5",
    "slowapi>=0.1.9",
    "sqlalchemy[postgresql-asyncpg]>=2.0.48",
]
